requirements.txt
# Existing dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.20
python-dotenv>=1.0.0
spacy>=3.5.0
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
PyPDF2>=3.0.0
beautifulsoup4>=4.11.0
nltk>=3.8.1
numpy>=1.24.0
pandas>=2.0.0
datasets>=2.14.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
tenacity>=8.2.0

transformers>=4.30.0       
torch>=2.0.0                
scikit-learn>=1.3.0         